Use location instead of quickfix list for errors
Since formatting works only on the current buffer, it could be argued that
formatting errors should be displayed in the current buffer window instead of
the global quickfix window. (Because the global quickfix window may already be
displaying other information, e.g. project-wide search results, that is lost
when formatting errors are loaded into it.)