Maintain folds after file is formatted
This commit fixes an issue where after the formatter is run, when the
file is reloaded, the created folds would all be unfolded.

Instead of calling `winsaveview`, which does not save fold
information, and using `winrestview`, to restore the view, this commit
opts to use `mkview` to save the current view, and then `loadview` to
load the saved view. This way folds won't be changed after the file is
formatted.